+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - When creating a new project, two digit dates are no longer accepted which
   means incorrect dates cannot be submitted.
+- The 'complete project' button is no longer shown on completed projects.
 
 ### Changed